Removed GPU instancing from improvements.
This was always an optional optimisation, and now I've realised drawing as separate tiles will work well if/when we support multiple lights in the future (cited as a requirement in #19).

From the docs: "When using forward rendering, Unity cannot efficiently instance objects that are affected by multiple lights. Only the base pass can make effective use of instancing, not the added passes."