Remove unnecessary placement scripts.
While working on something else i realised there is code in multiple places for adapting ocean to the current view (and they have separate references to the view transform).

This change consolidates them, reduces the code and removes update order dependencies.

The sphere offset is less interesting when the ocean shape contributes collision in the world (or dynamic sims), which are not related to the viewpoint.