Put ocean back in opaque queue, so it renders front to back.
Finally acting on this TODO - the ocean does not need to be in the transparent queue. It does not render with alpha blending enabled. This change puts it at the back of the Geometry queue and makes it render front to back which is much more optimal than back to front.