Change uses of ref to other modifiers in functions
The reason for this change is that ref implies that an input value
can both be read from and written to. However, for some parameters
they were only being read from (so were changed back to pass-by-value,
although the 'in' parameter modifier would be ideal in future versions
of C#).

Furthermore, some parameters were only being written to, so were changed
to `out` parameters.