Fix for ocean depth baseline
We need '0' to mean deep water. This is required so that when lods are faded in/out, it can fade to 0 (deep). It also means that outside the biggest lod the value can be 0 and not appear to be shallow. This fix changes the meaning of the depths - it is now the height of the sea floor above some deep value (currently -1000m).