#105 - Shader keywords not working in build - wow
Undo the change that made keywords driven by shaders. Shader variants not avail at runtime, unless ocean shader added to list of always include shaders, in which case the features generate a cool 0.5-1 million shader variations. So put it back to toggles on the material, but added warnings if feature not enabled to make it a bit less error prone.