Fix after pause glitch in waves
Coroutine was there to try to avoid the update function, but this could
get called multiple times per frame (for example when paused in editor)
which meant displacements could accumulate over time. I cant just
disable the script because it would deregister the draw. So leaving the
update function there.