Assign shader params as arrays to LodData
At the moment this is the point at which I hope the new system that
uses Texture Arrays will be in its *most* hacky state. It is very, very
hacky as-of this commit.

Although Animated and Dynamic waves at working as well as desired,
everything else seems to be in a fairly allright working order - with
the depths being rendered in a single drawcall using geometry shaders
to target multiple LODs.

A compute shader is used to combine the different animated wave LODs,
although this is producing weird pixelated results.

The goal now is to go through each LOD individually, and tidy them up so
that they are functional and use clean systems and APIs. This will be
done in conjuction with tackling TODOs.

The next phase will be to then port LODs to Compute/Geometry shaders as
appropriate. Then we will verify that the LODs are producing the same
results they did under the previous system.

Finally, getting the Ocean to render properly will be the order of the
day, whilst cleaning up the shader so that it more explicitly uses
texture arrays instead of the current workarounnd hacks.

Similarly, the LOD viewer needs to be cleaned as well so it doesn't
perform loads of copies.

The branch can the be kept up-to-date with master, and will be ready
for testing. Measuring draw-call reduction and performance gains will
be vital for evaluating if this change is worth merging. We can then
investigate the benefits of techniques such as async compute as well.