Implement filtering in ShapeCombine shader
If this commit seems hacky, it is because it is. It was made without
access to any HLSL documentation - meaning that bilinear filtering was
implemented by hand. However, I'm commiting it anyway as it does provide
nice verification that the AnimatedWaves LOD system does work with
TextureArrays! It probably has a couple of bugs, and this approach will
need to be changed. But we can focus on solving other more important
problems before we come back to this now.