Max detail height affected only by waves
Max detail height - camera height relative to sea level where ocean is
set to minimum scale to maximise detail - was affected a lot by
whirlpool. It caused it to lower which mean ocean was not at full
detail when camera was around sea level. Some lowering is ok -
it helps for gerstner waves which drop the average surface height
a lot - but the whirlpool should not lower it, so i've split out some
data so it can distuingish between gerstner displacements and other.