Add support for LodDataInput shader options
So far this only extends to allowing for LodDataInput to exist which
support geometry shaders, but this could be extended further to support
compute shaders (for example), or even static textures directly (through
simplified interface). For now making this limited change in order to
fix a bug due to the switch to sea floor depth geometry shader support.