Implement custom masks for underwater post proc
Motivation for this work:

A user on discord is working on replacing some Crest geometry with
their own but will still like their own underwater effect to be applied.

This change adds a simple script (UnderwaterEffectFilter) that can be
attached to any arbitrary game object with a Renderer.

For now, it simply re-renders that object with the shader mask and it
will be treated as regular ocean surface.

The plan is to make this system completely configurable, so that by
adding this "filter" to game objects, users will be able have masks
which can disable the underwater affect when inside objects, or apply
it from a specific depth in the scene (like a window in a submarine).