Warn on reuse of a SampleHeightHelper multiple times in a frame
This is a big pitfall that is too easy to get wrong - the intuitive
thing is to reuse SampleHeightHelper multiple times, but it will
simply overwrite the query position before the result is sent
to the GPU (if the compute height queries system is used).

This change throws a warning if the position is set multiple times
in a frame.

Unfortunately there is a case where this is valid - FixedUpdate can
be called multiple times per frame. I added an arg to indicate this
which will suppress the warning. Some collision providers will be
happy to receive multiple queries per frame. The compute height
query system will just overwrite the query pos and send off the
final (last) fixedupdate position to the GPU. This is imperfect but
I don't see how this can work better.