Remove OnWillRenderObject()
This is a proof of concept enabled by recent refactors. Each per-frame
function on the ocean tiles is called 50+ times per frame so good targets
for optimisation.

I'm not 100% confident this can be applied to BIRP as there is a bit of
code removed that might be required, but i think its simpler in SRP.

This also removes the code that changes the ocean material. This would
need some new way to change the ocean material at runtime, probably
event-driven rather than checked every frame.