Rise up verts near far plane to camera height - ensures water is at horizon.
Depending on the ocean scale there may not be any verts at all near the
far plane, so i also pull verts up based on lodAlpha at the last lod.
This can look a bit nuts, and it pops when the scales change which would
need love.

Dread to think of all the issues raising the surface up through
geometry/terrain will cause as well.