Prototype resolution override for lods
Seems to work but i feel like i cheated as it was too easy. It wont support
things that need to compute derivatives. But i dont think its needed for
albedo. @daleeidd lmk if you have thoughts.

The other issue is memory usage - having all cascades is probably
overkill. Not addresing that for now.