Fix shadow sim running for all cameras
CBs on lights will run for all cameras so we need to restrict it to a
single camera by adding and removing the CB using the camera events.

This broke XR spectator camera and could cause incorrect shadows in the
simulation.

There is a possible Unity bug where the shadow input draw calls are
duplicated and without any commands applied to them (including setting
the render target).