🐛 Fix water renderer bounds thrashing in edit mode
Fixes renderer bounds thrashing in edit mode if using RegisterAnimWavesInput,
RegisterHeightInput or Whirlpool.

There was a mismatch in reporting maximum displacement where the OR was clearing
reported displacement in EditorApplication.update but these components were
reporting in Update. Chunk renderer bounds are updated in Update so it would
have alternating reported displacements which caused thrashing. It incurred a
performance penalty and incorrect culling.