Reduce WaterBody material override leaks
The water tile's expanded bounds would cause materials to leak outside of the
water body.

Use the unexpanded bounds to reduce leaking as generally this feature will be
for an inland body of water where hopefully there is attenuation between it and
the ocean to handle the ocean's displacement. The inland water body will unlikely
have large displacement but can be mitigated with a decent buffer zone.