Fix #1127: Clip surface affecting extents
When working out the slices for clip surface data the maximum slice
prevented the off-LOD weight from being used causing no transition and a
clamped sample discarding the extents (resulting in popping). A bug from
moving to PosToSliceIndices;

Clip surface is now blended out when nearing the end of the LOD chain.